requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
jinja2==3.1.3
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.6
httpx==0.26.0
